refactor test mocks into a separate file and create a context manager
